Split the curtailment module along its layer boundary
It sat in $lib/facilities/ — the cross-route facility layer, whose other
modules have 2-17 consumers each — with a single consumer, and it carried
/api/network/data metric keys that the facility layer has no business
knowing.

- Move the scoping half to charts/facility/curtailment/scope.js, beside the
  only component that reads it. This matches the sibling pollution panel,
  which keeps npi-url.js and its other helpers co-located. Promote it back
  up if a second route ever needs it
- Move the split-to-metric mapping into network/market-metrics.js, next to
  the registry whose keys it has to agree with, so a rename becomes a
  same-file edit rather than a cross-package one
- Move the metric tests across with it, leaving scope.test.js covering
  scoping alone
